construction-platform/python-services/api/requirements.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.2.1

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.2

# Prometheus monitoring
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=7.0.0

# AI/ML
google-generativeai>=0.3.0
sentence-transformers>=2.2.0

# Vector DB
qdrant-client>=1.7.0
supabase>=2.0.0
pgvector>=0.2.0

# Database
redis>=5.0.0
psycopg2-binary>=2.9.0

# OCR for construction documents
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
PyMuPDF>=1.23.0
reportlab>=4.0.0

# Testing
pytest>=7.4.3
httpx>=0.25.1

# Phase 3 Improvements
qdrant-client>=1.7.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-jaeger-thrift>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
boto3>=1.34.0  # For S3 archival
psycopg2-binary>=2.9.9  # For PostgreSQL
sqlalchemy>=2.0.23  # For database ORM


# Phase 4 Improvements
sqlalchemy>=2.0.23  # Database ORM (already added in Phase 3)
locust>=2.17.0  # Load testing
pytest>=7.4.3  # Testing framework (already added)
pytest-asyncio>=0.21.1  # Async testing
pytest-cov>=4.1.0  # Coverage reporting
